Extracted side nav to own component
